dol2gci in C
This tool barely uses C++.  It might as well be made in C.
